email verification logic

+++ database/migrations/2017_07_31_092922_create_jobs_table.php
@@ -17,7 +17,7 @@ class CreateJobsTable extends Migration
             $table->bigIncrements('id');
             $table->string('queue');
             $table->longText('payload');
-            $table->tinyInteger('attempts')->unsigned();
+            $table->integer('attempts')->unsigned();
             $table->unsignedInteger('reserved_at')->nullable();
             $table->unsignedInteger('available_at');
             $table->unsignedInteger('created_at');
